+++ price-feeder/oracle/oracle_test.go
@@ -108,6 +108,7 @@ func (ots *OracleTestSuite) SetupSuite() {
 			},
 		},
 		time.Millisecond*100,
+		make(map[string]sdk.Dec),
 	)
 }
 
@@ -479,6 +480,7 @@ func TestSuccessGetComputedPricesCandles(t *testing.T) {
 		providerCandles,
 		make(provider.AggregatedProviderPrices, 1),
 		providerPair,
+		make(map[string]sdk.Dec),
 	)
 
 	require.NoError(t, err, "It should successfully get computed candle prices")
@@ -511,6 +513,7 @@ func TestSuccessGetComputedPricesTickers(t *testing.T) {
 		make(provider.AggregatedProviderCandles, 1),
 		providerPrices,
 		providerPair,
+		make(map[string]sdk.Dec),
 	)
 
 	require.NoError(t, err, "It should successfully get computed ticker prices")